requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.12.7
csv2pdf==0.1.4
defusedxml==0.7.1
fonttools==4.38.0
fpdf2==2.6.1
numpy==1.24.2
pandas==1.5.3
Pillow==9.4.0
python-dateutil==2.8.2
python-telegram-bot==13.15
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.1
tzdata==2022.7
tzlocal==4.2
